// src/api/generated/models/recommendationPreferencesModel.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Tg-Api 
 * OpenAPI spec version: 1.0.0
 */
import type { LanguageModel } from './languageModel';

export interface RecommendationPreferencesModel {
  allow_clt: boolean;
  allow_contractor: boolean;
  allow_internship: boolean;
  allow_on_site: boolean;
  allow_hybrid: boolean;
  allow_remote: boolean;
  on_site_job_radius: number;
  hybrid_jobs_radius: number;
  allow_stack_flexibility: boolean;
  /** @nullable */
  min_remuneration: number | null;
  /** @minItems 0 */
  blackListedLanguages: LanguageModel[];
  dev_profile_id: string;
}